float version

+++ Temperature.c
@@ -4,7 +4,7 @@
 当fahr = 0, 20, ··· , 300 时,分别打印华氏温度和摄氏温度对照表
 */
 main(){
-    int fahr, celsius;
+    double fahr, celsius;
     int lower, upper, step;
 
     lower = 0; // 温度表的下限
@@ -14,7 +14,7 @@ main(){
     fahr = lower;
     while(fahr <= upper){
         celsius = 5 * (fahr - 32) / 9;
-        printf("%d\t%d\n", fahr, celsius );
+        printf("%.0f\t%.1f\n", fahr, celsius );
         fahr = fahr + step;
     }
 }
